MCP-2043: added camel bean validation. Organized tests
