+++ app/src/test/java/gov/va/vro/controller/FetchClaimsTest.java
@@ -1,9 +1,10 @@
-package gov.va.vro;
+package gov.va.vro.controller;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import gov.va.vro.TestDataSupplier;
 import gov.va.vro.api.responses.FetchClaimsResponse;
 import gov.va.vro.persistence.model.ContentionEntity;
 import gov.va.vro.service.spi.services.FetchClaimsService;
