+++ service/provider/build.gradle
@@ -25,6 +25,7 @@ dependencies {
   implementation "org.springframework.boot:spring-boot-starter-data-redis"
   implementation "org.apache.camel.springboot:camel-spring-redis-starter:${camel_version}"
   implementation "org.apache.camel.springboot:camel-jsonpath-starter:${camel_version}"
+  implementation "org.apache.camel.springboot:camel-bean-validator-starter:${camel_version}"
 
   // Needed?
   // implementation "org.apache.camel.springboot:camel-rabbitmq-starter:${camel_version}"
