+++ service/provider/src/main/java/gov/va/vro/service/provider/camel/MasIntegrationRoutes.java
@@ -104,6 +104,7 @@ public class MasIntegrationRoutes extends RouteBuilder {
 
     from(ENDPOINT_MAS_PROCESSING)
         .routeId(routeId)
+        .to("bean-validator:payload-validator")
         .setProperty("diagnosticCode", simple("${body.diagnosticCode}"))
         .setProperty("veteranIcn", simple("${body.veteranIdentifiers.icn}"))
         .setProperty(
